Rust: Add pr_*_ratelimited
This commit adds equivalent pr_*_ratelimited to kernel crate.
They are macros that define a static local `ratelimit_state` and
uses Atomic and CAS to guard its initialization.
That `ratelimit_state` will be used later
by `___ratelimit()` to decide if to suppress the `printk` or not.

Because we use Atomic & CAS to help initialize `ratelimit_state`,
in concurrent situation some thread may reach `__ratelimit()`
before `ratelimit_state` is initialized. If that happens, we
call `printk` anyway.

This commit does not implement pr_debug_ratelimited since it's
special and way more complicated and deserves another dedicated
commit.

`printk_ratelimit` is also implemented in this commit

Signed-off-by: Fox Chen <foxhlchen@gmail.com>